[focus] Don't send data: URLs when loading error pages (https://github.com/mozilla-mobile/focus-android/issues/455)
WebView seems to be a bit broken when loading data: URLs when
there's no network connection. It sometimes returns the data: URL
(instead of the original page URL), this happens at the same time as
when it doesn't send onLoadFinished(). Unfortunately there doesn't
seem to be any good workaround other than filtering out the incorrect
URL when it happens. It's impossible to predict when this will happen.