[components] Issue https://github.com/mozilla-mobile/android-components/issues/174: Sample browser: Move "browser behavior" into standalone fragment.
This is in preparation for tab support where we will start to switch between multiple fragments (e.g. browser -> tabs tray).
Currently this fragment lives in the sample app. I was thinking to move (partial) functionality back into components.. but
this was too complex for now. I first want to see where this is going and then at the end see what parts can be moved to
a shared component.