[components] Use `filesDir` as the parent folder for the places database file
Using getDatabasePath doesn't work without extra work on older API versions.
That location doesn't exist, and it seems like we'd need to add a bit of code
to check for its presence, create it, etc., before we can rely on that location.

All of our other code which stores files just uses `filesDir`, and I don't see
a reason not to put `places.sqlite` into that location as well. There isn't much
value in that file living in a "databases" folder.