[components] Closes mozilla-mobile/android-components#1293: Allow configuring how module tasks are scheduled; parallelize samples-browser variants
This splits the singular :samples-browser task into tasks for most of its variants.
Module overrides are made configurable so that we may easily do the same treatment
of other modules.

Additionally, lint is switched to just running for a single variant of the :samples-browser.
Linting all the variants is unnecessary and adds a significant overhead.