[components] Part 5: Do not emit 'logged-out' event on "no account present"
The change here is to only emit the "logged out" event when a logout
action takes place. Before we were emitting it as part of the
"i'm initializing, and there is no account present" flow.
That complicates a sync manager implementation, as it'll have to
distinguish the two states. By dropping this event, we allow
consumers assume that initial state is always "logged out",
and the correct event observers will fire otherwise.