[components] 1539480: Copy the user-lifetime metrics from old to new location
If this isn't the first run, it looks for the two user-lifetime metrics in
the old location of `glean_ping_info` and copies them to `glean_client_info`.
It is not problematic that the old values remain in `glean_ping_info`, since
the contents of `ping_info` are hard-coded and don't include these values.