[components] Persist FxA account state via the persistence callback
This is the preferred way to make sure our persisted account state
is up-to-date.
In the future, various parts of FxA (like SendTab) may mutate
account's internal state, and so it's necessary to make sure we
persist account changes as they happen.

Before, this happened manually after account creation.
Now persistence process is entirely callback-driven.