[components] Part 1: Do not erase account storage after hitting an auth problem
Before this patch, we'd clear out account storage after we see an unrecoverable authentication error. In this context, a password change would be an unrecoverable error. We'll transition account into an AuthenticationProblem state, but upon re-initialization of the state machine we'll end up in NotAuthenticated instead, because there's no longer a locally persisted account state. So, this is just a straight-up bug in the state machine.

So, step 1:
    do not erase account storage if we encounter auth problems we can't auto-recover from
        this will make sure account isn't disappearing after a restart
        this has an additional benefit of not erasing fetched profile data, as well! it's part of the persisted account state

Also, add in some debug logging into the samples-sync app to make testing this a bit easier.