[components] for mozilla-mobile/android-components#5603 - Implement notification for when the migration is complete
Currently a notification is being shown for migration progress status, but no notification is being
displayed to the user for when the migration is complete.
In order to properly handle this notification's remove or click action, a broadcast receiver is
required that listens for this specific actions so it can remove the notification's channel.