[fenix] Issue https://github.com/mozilla-mobile/fenix/issues/8277: Reuse app icon in widget layout.
This patch refactors the widget layouts so that we can reuse the app icon and do not have to have
another logo resource with different constraints. This will also make it easier to update the
branding in https://github.com/mozilla-mobile/fenix/issues/8277 and https://github.com/mozilla-mobile/fenix/issues/8278.