[components] Closes mozilla-mobile/android-components#6297: GeckoEngineView: Detach SelectionActionDelegate if session is no longer rendered.
Our sample browser smoke tests revealed a memory leak (mozilla-mobile/android-components#6297). We attach a selection action delegate on
GeckoSession but we never clear it. This leads to us leaking GeckoEngineView and its Context/Activity
since the GeckoSession usually lives longer than any views.

Internally GeckoSession attaches and detaches its own selection delegate (since it also has a reference
to the activity). We are now doing the same if we are using a custom delegate.