[components] Closes mozilla-mobile/android-components#7180, (mozilla-mobile/android-components#7939): Create single download dialog and don't reuse once destroyed
Co-authored-by: Arturo Mejia <arturomejiamarmol@gmail.com>

With this we're no longer holding a reference to the download dialog in the
feature, which caused a memory leak, as we tried to reuse the dialog
instance after it was destroyed. In addition, we now also make sure we only
ever create a single instance of the dialog by relying on `showNow` to
prevent a delayed `commit` of the fragment transaction.