[components] Close mozilla-mobile/android-components#8531: Use synchronized block in PushConnection init check
This was the only call that did not have the initialisation check around
it. It may have been the cause for a race seen in Fenix send tab, but we
should fix this regardless to always ensure synchronization.

No tests are added since there isn't a consistent way to test a block is
synchronized.

Co-authored-by: Grisha Kruglov <gkruglov@mozilla.com>