[components] For mozilla-mobile/android-components#8565 - New RepostPrompt that inform if users don't proceed with reload
The prompt for a repost request was previously an instance of Confirm, type
used for many other requests.
We'll now use a new Repost type that allows easily identifying the particular
scenario for a repost request, prompt that will now show a slightly different
text and will notify engine observers of "onRepostPromptCancelled" when users
asked to refresh the page but then cancelled that operation in the repost
prompt, callback used to hide the pull to refresh throbber.