[components] Closes mozilla-mobile/android-components#8672: Sample-browser: Use SearchEngine from BrowserStore instead of SearchEngineManager.
This migrates sample-browser to stop using SearchEngineManager. For that I hide the provider of the default
search engine behind an interface. This is a temporary solution until we have all apps using the SearchEngine
from BrowserStore. After that we can get rid of that and make all our components use BrowserStore directly.

I tested this in Fenix too and have a matching patch locally. Once this lands we should be able to migrate
Reference Browser too.