[components] Closes mozilla-mobile/android-components#8715: Pass and empty map to loadUrl() instead of null.
GeckoView switched to a new `Loader`-based API and the old methods have been migrated under the hood to use that.
However it seems like this breaks when passing in `null` for the headers.

I filed a GeckoView bug for that:
https://bugzilla.mozilla.org/show_bug.cgi?id=1671571

I considered just switching to the new API (Issue mozilla-mobile/android-components#8710), but it's not straight-forward to write a unit test
for the `Loader`-based API. I filed another GeckoView issue for that:
https://bugzilla.mozilla.org/show_bug.cgi?id=1671578