[components] Close mozilla-mobile/android-components#8846: Re-subscribe FxA push subscription on subscriptionExpired
This is a workaround fix to re-enable Send Tab when we are notified
about the subscriptionExpired flag. Until (mozilla-mobile/android-components#7143) is fixed, we need to
avoid calling `registrationRenewal` to avoid going into an unfixable
state.

In the context of the FxaPushSupportFeature, we know that our end goal
is to call `push.subscribe(fxaPushScope)`, so we can by pass the
`verifyConnection` call **assuming** that our native push client is
always has the latest push token and knows how to invalidate the
endpoint with the Autopush server.