[components] Issue mozilla-mobile/android-components#8744: CustomSearchEnginesStorage: Use filesystem safe file names.
I saw crashes when testing the Fenix integration on a device (File not found when trying to read a custom search
engine from disk). We used base64 encoding, but forgot to avoid line breaks and to use file system safe characters
only.

Unfortunately this wasn't reproducible in a unit test. Apparently Robolectric is more forgiving then the actual
file system of an Android device.