[components] For mozilla-mobile/android-components#9189 - Refactor service-pocket to support recommended articles.
Since the previous version of this component was not in use anymore it was used
as a starting point for the newly needed functionality - easily offering Pocket
recommended articles to interested clients.

The new version offers a simple API for downloading, caching and refreshing in
the background the list of Pocket articles that is offered to clients.