[components] For mozilla-mobile/android-components#10470 - Add a new lastMediaAccess property for TabSessionState
This new property will be a timestamp of the last time media started playing in
the current page or be 0 if media hadn't started playing or another page loaded
in this tab so media was stopped.

To observe the media changes and update this property LastMediaAccessMiddleware
will have to be set on BrowserStore.