[fenix] For https://github.com/mozilla-mobile/fenix/issues/21138: change EXPECTED_RUNBLOCKING to range.
Compared with my local runs, CI sees +1 runBlockingIncrement calls so seems to
take other code paths. As such, we search for a range of rather than a single
value. The main downside I can think of is this could make the test trickier to
debug but we can wait and see if that's a problem in practice before taking
action on it.