[components] For mozilla-mobile/android-components#11753 - Update compose to 1.1.0 and Kotlin to 1.6.10
This also required updating room to >= 2.4.0.
This new version adds a deprecation of the `MigrationTestHelper` api used in
`LoginExceptionStorageTest` that is to be later fixed in (mozilla-mobile/android-components#11765).

activity_compose was also update to the latest stable version to ensure a
better match with the latest stable version for compose.