[components] Don't try to parse migration metrics from importLoginsAsync
A change in AS v93.6.0 made it so the underlying function
`LoginsStore.import_multiple` no longer returns the metrics as a JSON
string string.  Instead it just encodes the unit struct. This was a
breaking change, but it's not listed properly in the CHANGELOG.

The fix is pretty simple: just assume that no logins failed to import.