[components] Revert "Use the local server for testing"
This reverts commit 9f08e456b06a02556807664efc32d2fa99878cdb.

In Glean v50 `testGetValue` methods are actually not gated on test
mode (it's still not recommended though and absolutely IS still a
test-only API for all intents and purposes).

Further it turns out that calling `GleanTestLocalServer` in one test
might influence the other test and we're seeing intermittens where
`BaselinePingTest` is trying to upload to `localhost:0`, which of course
is invalid.
I have not yet figured out how we can properly separate out the state.
Alternatively we could merge the 2 tests into a single one.

Fixes mozilla-mobile/android-components#11340