[fenix] Close https://github.com/mozilla-mobile/fenix/issues/25919: Move UpdateFirstFrameDrawn to RecyclerView.onLayoutCompleted
As part of a preventitive measure for home page regression loading, we
fixed the UpdateFirstFrameDrawn call to happen after the first layout in
the main RecyclerView is completed. In addition, we also make pocket
aware of this flag so that it renders itself after the first layout.

This helps prioritize current & future features that are visible first
to render before those that are off screen.

Co-authored-by: Arturo Mejia <arturomejiamarmol@gmail.com>