Bug 1797245 - Revert mozilla-mobile/fenix-beta-version from 5.0.0 to 4.1.0
In v5.0.0, we assumed that we needed to get the version.txt from the project direction instead of the root direction. However, we will be switching the AC versioning to follow Focus. So, the root version.txt should be the same as the existing focus-android/version.txt going forward. To accommodate this change, we are downgrading back to v4.1.0.