Bug 1824434 - Improve visibility of license info  (#1501)
* Bug 1824434 - Improve visibility of license info

We are upstreaming some of the code that was vendored into Fenix, to be
shared from components, so that Focus can use it as well.

Why is this a new component? Inspecting our existing components, this
code doesn't fall under the concept of a "feature", and more closely
relates to being a "support" one instead. With that in mind, there
isn't a good existing support component to fold this into either.

There is also room for improvement in this component to include the
plugin or other license-related helpers.

* Bug 1824434 - Add UI tests for license menus

---------

Co-authored-by: dsmithpadilla <88508950+dsmithpadilla@users.noreply.github.com>