Bug 1826560 - consistently sign android-test APKs with the dep key
These APKs need to be signed with the same key as the ones used in firebase for
ui-test and startup-test, which already use dep signing.  Stop discriminating
between different tasks_for and just keep a list of production build types for
release signing.

Also for simplicity, remove the conditional for adding to the index since I
can't think of a reason this is necessary.