Bug 1831168 - rename built apk artifact to target.{abi}.apk
Instead of including the product name and abi as directories containing a
target.apk, remove the product name entirely and put the abi in the filename
directly.

The main motivation is to make the beetmover manifest simpler when adding
focus/klar because all signing tasks now have the same artifact names.