Bug 1831497 - [a11y] Dictate correct hint when selecting engine
In order for talkback to correctly dictate the hint for a newly
selected search engine in the search toolbar, it required setting
it when updating the content description of the toolbar.
Previously, when updating the content description, it was using
the previous search engine hint.

(cherry picked from commit 5edf51af11c2fa6f8a69465504f28a5ae545474d)