Bug 1672252 - Ignore case when appending Juno Onboarding privacy link
Due to a difference of lowercase/uppercase letters in the message
shown in Juno Onboarding when displaying the privacy notice link
and having talkback enabled, the app crashed because it could not
find the starting index of the required text. This patch will not
add the link text if the description text does not contain the link
text.

(cherry picked from commit e60f3b88cb8dd554f85cc843b1b70946a6da0c51)