Bug 1841749 - Prevent interaction with toolbar when elements on top
When toolbar is positioned at the bottom and the download prompt
appears, the user is still able to interact with the toolbar. This
issue is also reproducible when displaying the find in page bar or
the reader view controls, the user being able to swipe on these to
navigate to another tab.