Bug 1809828 - Display Tab Thumbnail when failing to download image
When trying to download an image provided by a website to set as
thumbnail for Jump Back In section of the home page, the download
might fail due to image being to large. In order to fix this, as a
fallback when the download fails, a snapshot of the website will
be provided as Tab Thumbnail. Previously, the fallback was a blank
white rectangle.