Bug 1850987 - Stop sending RecordNotFoundException to crash reporter
This is not an actionable error and if we need to know more details
about how often this happens, we should add telemetry instead in a more
meaningful way that would trigger some response.

Right now, no one really seems to look at this data and we are using up
our Sentry event quota unnecessarily.