Bug 1848842 - Increase APK regression alert's threshold to 1 MB
We added a regression alert for APK size regressions greater than 100 KB (in bug 1838717). This 100 KB threshold is too sensitive, so instead of disabling it entirely, I'd like to increase the threshold to 1 MB, so we will at least catch big surprises like -O2 bug 1838709 (+3 MB) and ICUX bug 1848025 (+2 MB).