Bug 1807147 - Query FIP results if string longer than one char
Currently the Find in page results do not work as expected when
searching for one character. This change finds results only
when the string searched for is at least two chars long. Another
issue here is that when `SessionFinder` results are invoked
in a quick succession, the number of matches is 0.

There is already a GV ticket here:
https://bugzilla.mozilla.org/show_bug.cgi?id=1587856