Bug 1855942 - Create compose version of `CFRPopupFullscreenLayout`
The current CFRPopup requires an anchor as parameter in order to
find the right position for placing the CFR. This current
implementation will not work on compose exclusive views due to
the lack of xml views to be used as anchor. This patch aims to
introduce a new CFR which can be displayed in compose layouts without
requiring a xml view.