Bug 1863754 - Disable sponsored and non-sponsored suggestions when Firefox Suggest is disabled.
This commit fixes an issue where it was possible for Firefox to show
sponsored and non-sponsored suggestions if the user explicitly disabled
and re-enabled the settings, and then unenrolled from the experiment.
In that case, the `show{Non}SponsoredSuggestions` settings would still
be `true` because they were user-set, even though `enableFxSuggest`
was now `false`.

As of this commit, the `show{Non}SponsoredSuggestions` settings are
only used if `enableFxSuggest` is `true`.