Bug 1857021 – add account deleted snackbar after custom tab account deletion
When the user deleted the account through the custom tab, it's not possible
to show a snackbar on that screen, as it's a separate activity, and snackbar
won't survive the host activity finishing. Added a flag to mark the need
for the snackbar to be shown, and a check for home and browser activity,
places where user is navigated in case of account deletion.