Bug 1857850 - [a11y] Improve experience when accessing links in text
This patch aims to improve a11y experience when accessing links
present in bodies of text. When we have a link text element, we
now indicate that there are links present in it and upon double
tapping, a dialog is open from where the user can access any of
the links present in the text. This behavior tries to follow the
default Talkback a11y one where the user can access links from a
text in case they are annotated as URLs.
This solution allows multiple links in one body of text and also
improves accessibility for lower version devices.