Bug 1872235 - Add lint detector for missing Response#close
In bug 1839067, we fixed a multiple cases were streams were
potentially leaking resources. As a follow-up, we are adding a lint
check that now helps to identify possible missing calls to close those
resources.

The implementation is heavily relied on from
[`com.android.tools.lint.checks.CleanupDetector`][0] in the AOSP
project, so little cleanup is done to change the code from it's
original source if we need to reference it later.

Some other references that were quite helpful for this work, were the
[Android Lint API Guide][1] and the examples in the [Data Flow
Analyzer][2].

[0]: https://cs.android.com/android-studio/platform/tools/base/+/mirror-goog-studio-main:lint/libs/lint-checks/src/main/java/com/android/tools/lint/checks/CleanupDetector.kt;l=66;drc=042516205eef87810ffe9f6d51ba40b1edbb15b0
[1]: https://googlesamples.github.io/android-custom-lint-rules/api-guide.html
[2]: https://android.googlesource.com/platform/tools/base/+/HEAD/lint/docs/api-guide/dataflow-analyzer.md.html